add circulo.cpp

+++ Tarea1-1.cpp
@@ -3,11 +3,9 @@ using namespace std;
 const float IVA = 1.19;
 
 int main( ){
-    int precio;
+    float precio;
     cout << "Ingrese precio" << endl;
     cin >> precio;
     cout << "Precio final: " << precio * IVA << endl;
-
-
     return 0; 
 }
